refactor: organize file paths
