+++ EShopCores/AppLogs/LogHelpers/ILogHelper.cs
@@ -1,11 +1,10 @@
-﻿using EShopCores.Enums;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace EShopCores.LogHelpers
+namespace EShopCores.AppLogs.LogHelpers
 {
     /// <summary>
     /// LogHelper Interface
