+++ EShopCores/AppLogs/LogModel.cs
@@ -1,11 +1,10 @@
-﻿using EShopCores.Enums;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace EShopCores.Models
+namespace EShopCores.AppLogs
 {
     /// <summary>
     /// 紀錄Log用的Model
