+++ EShopCores/EShopCores.csproj
@@ -8,6 +8,7 @@
 
   <ItemGroup>
     <Folder Include="Entities\" />
+    <Folder Include="Enums\" />
   </ItemGroup>
 
   <ItemGroup>
