+++ EShopCores/Responses/ResponseMessageAttribute.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace EShopCores.Errors
+namespace EShopCores.Responses
 {
     /// <summary>
     /// 自訂Response訊息屬性
